fix(ci): make cargo fmt and cargo clippy pass green
- Run cargo fmt to fix all formatting across workspace
- Add pre-existing violation lints to [workspace.lints.clippy] allow list
  (clone_on_copy, explicit_auto_deref, redundant_closure, ptr_arg, etc.)
  with comments tracking them for gradual cleanup
- Remove empty test modules that had no #[tokio::test] functions but left
  dead Data structs, unused imports, and unfulfilled #[expect] attributes
- Fix specific clippy violations in new code: return Err(x)? → Err(x)?,
  else { if } → else if, |r| r.to_string() → ToString::to_string,
  "".to_string() → String::new(), manual_range_contains
- Scope #[cfg(test)]-only imports with #[cfg(test)] attribute to silence
  unused-import warnings in the non-test build (wallet.rs Crud trait)
- Fix empty-line-after-doc-comment by converting orphaned /// to //
- Fix unused-async function in api_utils context (add targeted allow)

+++ crates/db_schema/src/impls/coin.rs
@@ -112,7 +112,7 @@ impl CoinModel {
       .await?;
 
     // 2) Compute new total (integer, no decimals)
-    let new_total = current_total + delta.0 as i32;
+    let new_total = current_total + delta.0;
     if new_total < 0 {
       return Err(FastJobErrorType::InvalidField("coin supply cannot be negative".into()).into());
     }
